+++ richat/src/bin/richat.rs
@@ -18,8 +18,8 @@ use {
 #[derive(Debug, Parser)]
 #[clap(author, version, about = "Richat App")]
 struct Args {
-    #[clap(short, long, default_value_t = String::from("config.json"))]
     /// Path to config
+    #[clap(short, long, default_value_t = String::from("config.json"))]
     pub config: String,
 
     /// Only check config and exit